+++ packages/@uppy/companion/src/server/provider/index.js
@@ -22,6 +22,7 @@ class Provider {
   constructor (options) {
     return this
   }
+
   /**
    *
    * @param {object} options