+++ packages/@uppy/utils/src/getTimeStamp.js
@@ -1,6 +1,6 @@
 /**
  * Returns a timestamp in the format of `hours:minutes:seconds`
-*/
+ */
 module.exports = function getTimeStamp () {
   var date = new Date()
   var hours = pad(date.getHours().toString())
@@ -11,7 +11,7 @@ module.exports = function getTimeStamp () {
 
 /**
  * Adds zero to strings shorter than two characters
-*/
+ */
 function pad (str) {
   return str.length !== 2 ? 0 + str : str
 }